+++ public/dialog_language_mode.js
@@ -2,7 +2,7 @@ import {ChangeEditorOptionEvent} from '/change_editor_option_event.js';
 import {DialogSelect} from '/dialog_select.js';
 import {Res} from '/res.js';
 
-export class DialogMode extends DialogSelect {
+export class DialogLanguageMode extends DialogSelect {
 
     constructor(core) {
         super(core);
@@ -17,7 +17,7 @@ export class DialogMode extends DialogSelect {
 
     open() {
         const res = new Res();
-        super.open(res.titles.mode, this.getListData());
+        super.open(res.titles.language_mode, this.getListData());
     }
 
     reset() {